Removed listing policy results module and moved its model to policies module
