--- airbnb_project/listing_policy_results/admin.py
@@ -1,3 +1,0 @@
-from django.contrib import admin
-
-# Register your models here.
