--- airbnb_project/listing_policy_results/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class ListingPolicyResultsConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'listing_policy_results'
