+++ airbnb_project/policies/models/listing_policy_result_models.py
@@ -1,6 +1,6 @@
 from django.db import models
 from airbnb_project.listings.listing_models import Listing
-from airbnb_project.policies.policy_models import Policy
+from .policy_models import Policy
 
 class ListingPolicyResult(models.Model):
   listing = models.ForeignKey(Listing, on_delete=models.CASCADE)
